+++ doc/requerimientos.md
